Use immediate when deleting user
I don't see harm in allowing other processes to read the user data while
the delete transaction is busy.  There's the potential danger that
another process starts a task that has been deleted already in the
transaction, but this was already not working with our Postgres
exclusive transaction, so it needs an alternate solution.