Lock table for Postgres when deleting report
The issue here is that exclusivity is needed to prevent other processes
from resuming the report's task between the time we check that the report
is not active and the time that we commit the deleted report.

Locking the table, instead of trying to lock the entire database, seems
like the best option at the moment.