Use lockfiles to control access at startup
Use a lockfile for each type of process instead of using exclusive
locks, to prevent certain types of processes from running
simultaneously.

The three types of processes are the main process, option (helper)
processes and the --migrate process.

Use a fourth lock, gvm-checking, to prevent races when checking the
three lock types, and also to allow the main process exclusive access
when checking the db initially.