Check for resources in use when deleting users.
When deleting users, check if resources are still in use by other users
after deleting resources that could use them. In this case, deleting the
user will fail.
The order in which resources owned by the deleted users are deleted has
been changed to ensure the referencing ones are deleted before the
referenced ones.