Skip credentials on all errors in slave_setup
The function omp_read_create_response used in the functions used to
create credentials returns -1 if there is no id in the response, which
effectively means it will only return the OMP status on success.
Because of this it makes no sense to expect an OMP status code from
the credential creation functions in slave_setup, so now credentials
be skipped on any error when trying to create them on the slave.