Test credential formats, show usable ones in OMP
A new OMP element has been added to get_credentials to indicate which
export formats will work for the type and login name of the credential.

The credential package generators will now also check if they accept the
credential data before attempting to generate the package.
If they fail for a different reason, a warning is logged.