Limit resources shown by get_tags, update GMP doc
The number of resources shown in the details of a tag is now limited and
the resources_first and resources_rows attributes have been added to
control which and how many resources to list.
By default the number is 100, which can be controlled with the new
setting "Tag Resources Per Page".

The GMP documentation for get_tags has also been updated accordingly
and to consider some changes related to multiple resources per tag
previously not included.