Use existing library functions for hosts_contains
In the hosts_contains SQL function use a new function that uses the
existing functions openvas_hosts_new_with_max and
openvs_host_in_hosts to avoid needing a new libraries function.
The number of hosts in the hosts string will also be limited to the
current max_hosts value.

Getting max_hosts in the PostgreSQL version is also moved to a new
static function for reusability.