Add index reports_by_task
This speeds up the slow task iterator that runs in stop_active_tasks
on startup.

An alternate solution would be to create the iterator by hand in
stop_active_tasks, as the iterator only uses the run status.  But
adding the index is good enough, and it may also help GET_TASKS
get the previous report faster.