Add Super preselect to ownership clause for permissions
This adds the Super preselection to acl_where_owned_user in the special case that
the type is 'permission'.  As a result it must also update the GET iterator init
to handle the preselect (i.e. to handle a WITH clause) because the permission
case now always returns a WITH clause.