Support two WITH clauses in init_get_iterator2_pre
This enables init_get_iterator2_pre to combine the WITH clause that it
receives from init_result_get_iterator_severity and the WITH clause that
it receives from acl_where_owned_with.

This means that init_get_iterator2_pre now always gets a WITH clause from
acl_where_owned_with, which means there is only one case left that is
using the non-WITH case of acl_where_owned_with: the result_overrides
view.