Prevent getting results and reports from trashcan
Prevent users from calling GET_REPORTS and GET_RESULTS with the trash=1
option.  This simplifies Manager internals, because we do not need to
add special cases to access the results_trash table when a result is in the
trashcan.  GSA does not use these options anyway, because it is the
associated task that is displayed on the trashcan page.