Build everything directly into the executable
This removes the sub-libraries (manage, gmp and otp).  Instead all the
source is included in single build target.

The sub-libraries were built because back in 2009 the idea was that
other programs could access the Manager data directly via libraries
instead of being required talk OMP (now GMP) to Manager.  This
never happened and it is not such a practical idea, so better to
remove the libraries and have a simpler build file.