Deactivate header files for doc-build.
Currently many header files produce warning for "make doc"
because it is required to document function protos. However,
these are already documented in the implementation part.

So, this commit deactivates the consideration by replacing
the "/**" by a "/*" in the "@file" block. This block was
added with an earlier commit of the same pull request indending
to be consistent. So, things don't get worse with this commit,
just not as good as wished for the long-term.