Remove smb_file_path_type alert data type.
Whether the smb_file_path is handled as a directory is now based on
whether the path ends in a slash, making smb_file_path_type obsolete.
The file extension from the report format is now always added to the
path, even if it is a filename.
The existing CHANGES entries have been updated accordingly.