Update CSV Results for new references.
This involves a slight change of the CSV format:
The columns for CVEs, BIDs, CERTs and Other References used
a comma as a separator inside the cell, Not instead of a comma
a newline is used as a separator.
The comma was a left-over convenience from the old references format.