Request nvti_cache update only at very end of NVT update
The nvti_cache is Manager's internal memory cache of NVT info. Previously the NVT
update process was flagging the nvti_cache for update in set_nvts_feed_version.
However, this happens at the end of the OTP plugins, before the OTP preferences
are received. This meant that the main process could update its nvti_cache
after the NVT update process had cleared the nvt_preferences table but before
the process had filled the table with the new preferences.

The result was an empty nvti_cache in the main process, which was causing scan
results to be recorded with incorrect information, like all having a 75% QOD.

Now the nvti_cache is flagged for update after the nvt_preferences table has been
updated.