Activate the use of nvti for sql insert.
The function insert_nvt is change to take just a
single paramter, the nvto object.
It will colect all required information from this
object for inserting it into the database.

Essentially this prepars the possibility to
have complicated list obejcts like references
considered by this function as well. With one-param-per-element
this would make the function call very complicated.

Another side effect is that it removes the bruden to handle all
the content of a nvti separately. make_nvt_from_nvti() is
simplified as it does not need to handle references on its
own anymore.
Also the functions get_ref(), get_cve(), get_bid() and get_xref()
are not required anymore and thus removed.

update_nvt_from_vt() is removed because the function from
the previous commit to create a nvti from XML is now actively
used.

The actual insert_nvt() calls now happen in get_vts_response.
This is also where the call of update_preferences_from_vt()
moved to from  update_nvt_from_vt().