Clean special option keywords in filters
When splitting a filter, some special option filter keywords
(like "first", "rows", "apply_overrides", etc.) will be changed so the
relation is changed to "=".
Also, numeric option keywords will be adjusted so their values are in
the valid range (e.g. "first" is set to 1 if it is 0 or less).