Remove manage_complete_nvt_cache_update().
This function is not needed anymore. Subsequently
also functions update_old_config_preferences(),
insert_nvts_list() and insert_nvt_from_nvti().

The function make_nvt_from_nvti() is actually not needed
also. But it might be used for chunked insertion of
nvts and therefore it is kept and a dumy call added to make
the compiler accept the presence.