Add a more complex test, for gmp_tickets.c
This adds a test for create_ticket_run, which is complex because
create_ticket_run calls into other parts of the gvmd code (copy_ticket,
ticket_uuid, log_event, etc).

The test creates 'mock' versions of those functions.

A gcc attribute is used as an easy way to override the mocked functions.
The alternative would be to link against less of the gvmd code and to
fill the gaps with stub functions.  But our code is not very cleanly
separated to this is a lot of work.  Should get easier as we clean
things up, if the attribute is not suitable.