Don't consider the case of nvt = '0' in queries.
Since the migrator ensured there is no such result with nvt = '0'
anymore, we do not need to consider it any sort of SQL query.
Even if is added a new result for some strange reason that has
a '0' as nvt oid, then simply this '0' is handled like any other OID.
It is actually a semantic issue, not a syntax issue anymore.