Add tls_certificates sha256 and serial, drop trash
This adds sha256_fingerprint and serial as table columns and params for
get_certificate_info and also adds an incomplete migrator that drops the
now unused tls_certificates_trash table and adds the new columns.

The migrator will fail intentionally in this commit because more columns
are to be added in following commits.