When adding a nvt preference, replace if exists.
Previously, when adding a prefernce that already exists
aborts the addition and keeps the old preference.

This is changed: Now if the preference already exists,
it will be deleted and then the new record be inserted.

It might be even easier to always deleted the record
without checking first for its existence.

Alternatively, if already exists, we could execute
and UPDATE for this record and only for new ones
an INSERT.

However, this patch works and no considerable
delays were measured  when replacing all existing
preferences compared to before.