When inserting new nvt, then replace if exists.
Previously, if a nvt is inserted which already exists,
the old one was kept.

Now, if the same oid already exists, then the nvt with this oid
will be deleted and afterwards the new one will be added.

The prior check for existence could be removed in favor of simply
always deleting the record, no matter whether it exists or not.

Alternatively an UPDATE could be executed instead of an INSERT.
But measurements have not shown a significant slow-down with the
present solution.