Handle solution and solutuion_type explicite in db
This prepares to have both not anymore in the pipe-separated tag list.
However, this patch makes the solution appear twice in the nvts table
for the time being. Just like solution_type already.

This patch expects that there is a additional column in the nvt
table "solution." A migrator is forthcoming, for testing this can
be done:
ALTER TABLE IF EXISTS nvts ADD COLUMN solution text;