Prefer uid over name for NVTs in manage_read_info
This gets RAW_DATA working for GET_INFO when type is 'nvt'.
handle_get_info is calling manage_read_info with both name and uid to
fill the RAW_DATA, so it got no data because manage_read_info was trying
to find the NVT using the NVT name.

There's no change in behaviour of GET_INFO when given an NVT as name or
info_id, because it is forbidden to give both.