get_nvts: Append solution/solution_type to tag
This also serves get_info for type NVT.
The explicit elements "solution" and "solution_type" are appended
to the pipe-separated tag element of the nvt.
This is a temporary solution until the XML structure of a NVT
has explicit solution and solution_type elements. Once
that is established, we don't need to append them to the tags anymore.