Remove forked == 2 checks
forked is never 2 anymore, as we no longer have the current_error
hack that was used to prevent the scan handlers from forking
again.  With OTP gone the forked scan handlers never return to the
main loop, and they make their own scanner connections, so these
checks are not required.

This patch is much clearer when ignoring whitespace.