Eliminate remains of get_tag().
The function get_tag() only parses the cvss_base_vector
from the incoming OSP XML for a VT and adds it to the nvti tags
string.

Using the new function nvt_add_tag() makes avoids the knowledge
about the internal syntax (pipe-separated list).

This last tag that is parsed is now located in function nvti_from_vt().

Also this removes the parsing for "cvss_base" inside the tags
string which happened in nvti_from_vt(). Actually it is not being
parsed anyway but rather computed from the vector.