Extend NVT element for GET_INFO
This adds the elements "summary", "insight", "affected"
and "impact" to the "nvt" element right after name and
before tags.

These elements only occur when data is found, else
they are not added.

It is the data that also occurs in the tags. So it is
redundant until the occurance in the tags element is finally
removed.