get_info nvt: Don't add solution elements to tags.
Since the solution and solution_type are now in a element
of its own, we do not need them in the tags element of the nvt
for the get_info response.

This change should be merged after GSA parses the solution and
solution_type properly from the new elemnt for the get_info command.