Change rows of built-in default filters to -2
If no filter or a filter without a rows keyword is given, the various
GET_... commands using filters will behave as if "rows=-2" was given
instead of "rows=-1".

This means they will only return the number of rows defined by the
"Rows Per Page" setting and the maximum amount of rows has to be
requested explicitly with a "rows=-1" filter keyword or
(where available) the ignore_pagination option.