Remove termination_signal handling from handle_slave_task
The variable is no longer set, because the signal handlers now simply
exit the process.

This means that if a slave handler process is terminated the associated
task will no longer be set to Stopped.  But we already had this problem
if someone SIGKILLs the process so it's not too bad.  Also, this is
consistent with the handlers for other types of tasks.