Prepare sync scripts for combining them into one
This changes the CERT and SCAP sync scripts so they are essentially the
same except for the feed names, URLs and paths.
Also, some initializations have been moved around so they can be run
after parsing the arguments so a type argument can be added to the
generic script later.