Add config preferences even if they are not in nvt_preferences
This ensures that the config preference is created even when the NVT or NVT
preference has not arrived in the feed yet.  It is required because syncing
of configs no longer requires the NVT preferences to all be present (in
order to speed up availability of feed defined configs during initial sync).