Improve EXE installer NSIS script generation
The temporary VB script and output file for getting the admin group name
are now created with random filenames.
Also, the user is now created and added to the group directly in the
installer instead of generating and running a batch script and the
command containing the password is hidden.
Temporary files are also written and read by the installer instead of
using echo commands for better readability.