Remove DROP from vulns creation
This DROP was added in 01baac22a, which changed the view to a UNION with
explicit columns.  It's likely that the DROP was needed because CREATE OR
REPLACE does not support changing the columns.

But that was 2017 and the columns have not changed since, so anyone
upgrading from the last version will have the same columns.

If someone need to change the columns they can add a migrator.