Update GMP doc for new severities and score
This moves the definition of the new common `severities` element to
a reusable element type definition and adds it to various responses
where it was missing.

The `get_info` types where `cvss` and `max_cvss` have been replaced
by `score` are also changed accordingly and a change notice has been
added about this.

Various command examples are updated to reflect the above changes.