Use timestamp in automatic sensor task names
The automatically generated task names for tasks on GMP sensors now also
contain the date and time in ISO format.
This is meant to avoid naming conflicts of the related targets, configs
etc. if the task is started again after an error where the resources
created by the master were not removed or if two tasks with the same
name are run on the same sensor.

(cherry picked from commit 7a885911fc3f19cdb41b96def4e41c2e98a2a178)

# Conflicts:
#	src/manage.c