add <detection></detection> on report import
When importing a report with <detection> and even if it has the correct
nvt as a host_detail as well as an result within the report the
detection block was missing when importing via create_report.

To add those sections two new entries within host_details will be added
when a <detection> got found within a result.