Refactor modify_config, allowing multiple changes
The modify_config command is rewritten using the new parsing method
creating an entity_t of the whole command XML and then processing it.
This also moves most of the GMP functionality of the command to
gmp_configs.c and allows multiple changes like modifying the name,
preferences and nvt_selections at the same time.