Also create owner WITH clause for single resources
The function init_get_iterator2_with will now also use acl_where_owned
to generate a WITH clause in case permissions_subject is used in a
subquery.
The init_result_get_iterator_severity had to be adjusted to not generate
the same WITH clause twice.

This change fixes an error getting details pages for types use the
WITH clauses for permission checks.