Fix report host end time check in CVE scans
When looking up the latest finished report_host for a CVE scan, the
query will also consider the end_time being set to 0 instead of NULL
when the scan of the host is unfinished.

Without this the CVE scan could return no or less results for hosts
being scanned at the moment.