Add --feed-lock-timeout option
This new option sets a number of seconds to retry for if the feed is
locked in contexts (like migration or rebuilds) that do not retry on
their own (like automatic syncs).
This is intended to help automated upgrades of GVM modules where
ospd-openvas needs to be restarted, which locks the feed while the
VTs are reloaded.

(cherry picked from commit 8e080ee3312fb2acf8d38b20a6b9fe6ca4d8367d)