Fix whole-only config family selection
When modifying the family selection of a scan config, "whole-only" must
be either growing with all VTs or static and empty.
The error message if this is not the case now also includes the name of
the (first) rejected family.

(cherry picked from commit 4d551cdd3019c7e6cb837003c2eae6337d312b1f)