Use less report cache SQL when adding results
When multiple results are added when handling an OSP get_scan response
or a host in a CVE scan, only one SQL statement each is run to update
the report and owner of the results and to update the end times of the
report_counts cache of the report.

This addresses AP-1495.

(cherry picked from commit 6d1797a866f89c7efe19edb04c5cc99edb5002c7)

# Conflicts:
#	src/manage_sql.c