Sort the "host" column by IPv4 address if possible
When sorting results and other resources by the "host" field, the items
with IPv4 addresses are sorted first and numerically instead of using
the normal alphanumeric sorting.
The alphanumeric sorting differs if the numbers have different amounts
of digits, e.g. 192.168.0.100 would be sorted before 192.168.0.3 when it
should be the other way.