Database adjustment for the removal of type, scanner from configs.
Since the fields type and scanner of the database table configs
were only used in connection with OSP scanners and OSP scanners
are no longer supported, the corresponding code has been removed
from the gvmd-sources. This is the corresponding database
adjustment, where, amongs others, the fields type and scanner
are dropped from the table configs.