Removed the "#ifndef NDEBUG" to get a backtrace in gvmd-journal.
The backtrace functionality in gvmd.c was, because of the
surrounding "#ifndef NDEBUG" statement, only activated on
development systems. Now that statement is removed to also
get the backtrace (eventually listed in the gvmd-journal)
on GOS-systems of the customer.

(cherry picked from commit a02850169c5a53b734335beaa2f54cfb43b18d53)