Fix iCal next time calculation, use new SQL func
The iCalendar helper functions and next_time_ical SQL extension function
now have a new parameter for the reference time, which should be the
current time.
This allows using a consistent reference time throughout queries that
take a longer time to run.
The functions now no longer return the current/reference time as the
next time when it is the same as a recurrence time, which could cause
unwanted effects like scheduled tasks being started twice for the same
scheduled time.