Clean up old SQL functions now included in pg-gvm
If the extension pg-gvm is not installed yet, gvmd will try to remove
the functions now included in the extension before installing the
extension.
This is needed because with PostgreSQL 13.8 the extension can no longer
simply overwrite the functions.

(cherry picked from commit 1937eec61cff30949494838e342cce2dac7b4aa2)