Fix: buffer overrun in buffer_aggregate_xml() (#1880)
When compiling with gcc on 32 bit armhf:

    sizeof(double) == 8, sizeof(double *) == 4

If needing a buffer for double, allocating space for a pointer
is insufficient and leads to a buffer overrun when assigning a value.

We shouldn't confound long int and long int* either.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>
(cherry picked from commit 501700b50c330fb124cbe46e5c002aa2d2706333)