Fix: The number of results in triggered alerts. (#1895)
* Fixed the number of results in triggered alerts.

Now the number of results for automatically triggered alerts and
manually triggered alerts are the same.

* Fixed small memory bug.

* Use selected report for manually triggered alert.

Now the selected report is used for the manually triggered alert.
Previously only the latest report (not the selected one) was used
for the alert.

* Refactoring of the first solution.

* Added the header comment for the function "init_alert_get_data".

(cherry picked from commit ea67ea30227cae36d8252e3076db33948a5939ff)