Add: New options to get all hosts with a given OS
This adds new filter keywords for the get_assets GMP command to find
all hosts with a given OS, not just the ones where the OS is the best
match.
Another new filter column is also added specifically for the CPE-ID of
the best matching OS of a host.
Also, a new all_hosts element and filter column are added to OS
assets containing the number of all hosts with the OS.

This allows finding out where OS assets are still in use by hosts and
making the filtered hosts pages more and counts shown in GSA more
consistent.