Fix: Fix running `gvmd --migrate`
Create lock function based files in the GVM_STATE_DIR and not in the
GVM_RUN_DIR. The GVM_RUN_DIR is only available when `gvmd` is actually
running. This is not the case when a database migration is needed. In
that case GVM_RUN_DIR does not exist and the `gvm-create-functions` lock
file can't be created. But that lock file is required to run `gvmd
--migrate` and without the migration `gvmd` wont start as a service.