Fix: Make matching for prefs migration stricter
When migrating NVT and config preferences use regexp matching to ensure
only ones where the substrings can be extracted are selected for the
update.

This addresses the migration failing because of preferences containing
a colon but not having the expected pattern.

(cherry picked from commit 976049d567f2b854c29ecfb04acef91418855ce4)

# Conflicts:
#	src/manage_migrators.c