Fix: Fix check_preference_names iterator
The missing new_pref_name column is added to the iterator for updating
preference names and the conditions for selecting preferences are made
stricter so must match the expected pattern.

The missing column caused a segfault if there are any preferences that
need to be updated.

(cherry picked from commit 0fb48e55902ffedeecf791495c13776c5ec96db5)