Change: Lock table for each auto-deleted report individually
The auto_delete_reports function no longer keeps the access exclusive
lock on the reports table for the entire loop over the reports, but
instead acquires and releases the lock for each report to delete.

This gives other processes, like ones deleting reports manually, a
better chance to acquire the lock on the reports table.