Add: Feed data objects can now be deprecated
The data objects in the feed (configs, port lists, report formats) can
now be deprecated, which prevents them from being imported unless they
already exist in the database. Existing data objects are will be marked
with a new XML element to indicate they are deprecated.

This will be used to deprecate feed data objects that are no longer
maintained.