Change: Improve table locking with timeout + retry
When trying to acquire the report table lock with retries, do not use
NOWAIT and sleep (), but instead use the lock_timeout setting in the
current transaction.
The timeout is adjusted to be below the deadlock_timeout and the number
of retries is raised accordingly.

This allows waiting processes to acquire the lock as soon as it becomes
available, reducing wait times for transactions blocked by fast other
transactions.