Change: Allow any inheritor deleting user from CLI
When deleting users using the `--delete-user` option, do not apply any
restrictions on which user can be selected as inheritor.

This gives users with command line access a way to circumvent the
restrictions of the delete_user GMP command and addresses issues
of the command line option being too restrictive to be useful.