Change: Optimize permission subqueries in acl_where_owned_user
The clauses checking if a permission for a given resource exists no
longer contain the resource id. Instead the id is looked up outside the
subquery in all permissions for the given type and user gathered by the
subquery.

This improves performance for resource types with many items like
results by allowing PostgreSQL to execute the subquery only once and
reuse it instead of having to run a subquery for every resource in the
database.