CI: Rework gvmd build container image
Change the gvmd-build container image build to be independent of changes
and releases of gvmd itself. The images only need to be rebuild if
gvm-libs has changed or if the base container image has new packages for
the dependencies. Therefore it isn't required to rebuild them on every
release, pull request or push.

Now only two flavors of the gvmd-build image exist. First
gvmd-build:stable which is based on the latest gvm-libs release and
second gvmd-build:edge which is based on the latest changes in
gvm-libs.