Change: Replace CPE xml_split with XML iterator
When updating the CPEs, the CPE dictionary file is no longer split
into separate temporary files. Instead a new XML file iterator added to
gvm-libs is used to avoid creating the DOM of the whole file.
Also, the buffered insert statements are now also run when a size
threshold configurable via a setting is reached.

This removes the dependency on the xml_split script, avoids problems
with the temp split XML files being left behind in case of crashes and
gives more control over the memory usage of the CPEs update.