Fix: Speed up CVE affected products update
When the affected products of CVEs are updated as part of the SCAP
data update, all newly added placeholder CPEs are now added to
the hashtable that is used to look up their row ids.
This avoids having to look up CPE row ids via SQL when inserting
entries into the affected_products table.
Also, to filter out duplicate product references, a hashmap is now used
instead of iterating over all the product XML elements.

This speeds up the SCAP update when there are CVEs with many CPE
references, especially if they do not appear in the CPE dictionary.
The most noticeable case for this is the initial update or rebuild
of the SCAP database.