Add: GET_FEATURES command and CVSS3_RATINGS toggle
The GMP command GET_FEATURES has been added to get the status of
optional features that can be enabled or disabled at build time.

This includes the OPENVASD toggle as well as a new one (CVSS3_RATINGS)
that will later be used to toggle CVSS3 / CVSS4 severity ratings.

The new command will allow GSA to get the current state of these feature
toggles.