Small fixes in adding Kerberos to targets
The not found error for the Kerberos credential in create_target now
uses the correct id.
For create_target the mutual exclusion of SMB and Kerberos is
checked earlier.
If the type check for the Kerberos credential in create_target or
modify_target fails, the type string is now freed.