Fix: Limit query size for affected products update
The update of the affected products is now limited to 1000
CVEs per SQL query. This number can be changed with the
--affected-products-query-size option.
Also, an index has been added to make the queries a little faster.

This is meant to prevent the WAL from becoming very large during the
affected products update.