Change: Handle timeout more like other VT prefs
The timeout preferences of VTs in configs/policies are now handled
more like normal VT preferences instead of a scanner ones, which is
also more consistent with default preferences of the VTs themselves.

This simplifies the special handling needed and fixes timeouts in the
config being overwritten by their default values.