Add scan queue handling
A new scan queue is added which can be enabled to limit the number of
scan handlers which are active at the same time by periodically
switching which tasks / reports are being updated.

This allows more tasks to be running at the same time without consuming
resources by having the scan handler run the whole time.