Remove delete_agent_installer GMP command
The agent installers can no longer be deleted by users via GMP
and the corresponding trash tables are no longer used.
Instead the feed sync will also remove any installers not in the
metadata file from the database.

This simplifies the file handling for agent installers and resolves
the issue that removing files from the feed would make installer
entries in the database unusable.
Now this has to be managed only by keeping the installer and metadata
files consistent with each other.