Fix: Add missing and amend G_LOG_DOMAIN #defines
Several G_LOG_DOMAIN #defines are added to non-test .c files where they
were previously missing and some existing ones are amended to be
consistent with the domains in the log configuration.

This fixes log messages not being filtered correctly, which could for
example result in unwanted debug messages with the default log
configuration.