Remove separate default openvasd scanner
The gvmd startup will no longer create a separate openvasd scanner,
but for new databases the "OpenVAS default" scanner will be of the
openvasd type if the OPENVASD build flag is set.
Actions that require the default scanner like checking the VTs feed
status will check the type of the default scanner instead of determining
the type of scanner to connect to based on the build flag.

This makes the behavior easier to understand by having only one default
OpenVAS based scanner and discourages the use of using both ospd-openvas
and openvasd on the same system, which may not work correctly.