Fix: handle 0 ports case in get_port_lists filter
The SQL for getting the port counts is adjusted to correctly handle
the cases where there are no TCP, UDP or any ports.

Without this fix, sorting using the "tcp", "udp" or "total" filter
column would return port lists with no ports of the selected type
in the wrong order.