Add `verify_credential_store` GMP command
The new GMP is added to check if the current host, path and preferences
of a credential store are valid and if the connection works.
A basic implementation for CyberArk without the connection test is
included as well.
The file manage_credential_stores.c missing from previous commits
is also added.