Add: New --secinfo-update-strategy option
The new command line option is added to give control over how the
SecInfo data is updated.
Currently the two options are the default of dropping the old
database schema at the end and dropping it at the start of the update.

The new non-default option is there to allow updating the SecInfo when
there is not enough database storage space for two versions of the
schema at the same time.