Fix: Allow missing deprecatedBy in CPEs sync
When updating the CPEs, the deprecatedBy array can now be missing or
empty even for deprecated CPEs. Also, the logging in case of
missing or invalid fields now identifies the affected CPE.

This fixes an issue where the latest CPEs feed could not be loaded
because of a deprecated CPEs that does not have a deprecatedBy list.