Add: New private_key_info in get_credentials
A new element has been added to the get_credentials response that
contains indicating that a private key has been set and containing
the key type and SHA-256 hash if the key is valid.
This makes it visible which private key is set in a credential without
revealing the private key itself.