Add: Functions using SQL prepared statement syntax
Functions are added that run SQL statements using the prepared statement
syntax with numbered parameters starting with a dollar sign, e.g "$1",
and passing the values to the SQL library function.
These are meant to replace the current SQL statement functions that
insert values directly into the SQL string in printf style.
The new method is more secure because it does not require strings to be
quoted explicitly.